refactor: Money 클래스를 사용하도록 코드를 수정

+++ src/main/java/wayc/backend/cart/application/dto/response/FindCartOptionGroupResponseDto.java
@@ -32,7 +32,7 @@ public class FindCartOptionGroupResponseDto {
                 .cartOption(
                         new FindCartOptionResponseDto(
                                 cartOptionGroup.getCartOption().getOptionName(),
-                                cartOptionGroup.getCartOption().getPrice()
+                                cartOptionGroup.getCartOption().getPrice().intValue()
                         )
                 ).build();
     }
